shoppie-master - Copy (2)/requirements.txt
flask==3.0.2
flask-cors==4.0.0
google-generativeai==0.3.0
python-dotenv==1.0.1
werkzeug==3.0.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
PyJWT==2.8.0
Faker==22.6.0 